definiendo tareas de admin 7

+++ frontend/lista.controllers.js
@@ -1,6 +1,7 @@
+import { modalControllers } from "./modal.js";
 import { listaServices } from "./servicios/lista_services.js";
 
-const nuevaLista = (username, created_at, id) => {
+const nuevaLista = (username, created_at, id, cantidad) => {
   const fechaCreacion = new Date(created_at);
   const fechaFormateada = fechaCreacion.toLocaleString();
 
@@ -15,6 +16,7 @@ const nuevaLista = (username, created_at, id) => {
                     <tr>
                         <th>Usuario</th>
                         <th>Creación</th>
+                        <th>Cant.prod/th>
                         <th>Acciones</th>
                     </tr>
                 </thead>
@@ -22,6 +24,7 @@ const nuevaLista = (username, created_at, id) => {
                     <tr>
                         <td>${username}</td>
                         <td>${fechaFormateada}</td>
+                        <td>${cantidad}</td>
                         <td><button type="button" class="btn btn-danger" data-userid="${id}" >Eliminar</button></td>
                     </tr>
                 </tbody>
@@ -36,14 +39,20 @@ const nuevaLista = (username, created_at, id) => {
   card.querySelector("button").addEventListener("click", async (event) => {
     event.preventDefault();
     const userId = event.target.dataset.userid;
-    try {
-      await listaServices.eliminarUser(userId);
-      // Eliminar la fila de la tabla después de eliminar el usuario
-      card.remove();
-    } catch (error) {
-      console.error(error);
+
+    // Mostrar una alerta para confirmar la eliminación
+    const confirmacion = confirm("¿Estás seguro de que quieres eliminar esta tarjeta?");
+
+    if (confirmacion) {
+        try {
+            await listaServices.eliminarUser(userId);
+            // Eliminar la fila de la tabla después de eliminar el usuario
+            card.remove();
+        } catch (error) {
+            console.error(error);
+        }
     }
-  });
+});
 
   return card;
 };
@@ -57,6 +66,7 @@ const renderLista = async () => {
   try {
     const respuesta = await listaServices.listaUsers();
     const lista = respuesta.listado; // Acceder al arreglo de usuarios
+
 
     lista.forEach((elemento) => {
       tabla.appendChild(
@@ -75,3 +85,28 @@ const renderLista = async () => {
 export const listaControllers = {
   renderLista
 };
+
+
+const card = document.createElement("div");
+card.classList.add("mx-auto", "mt-4");
+
+const contenido = `
+  <div class="row">
+      <div class="col-md-12">
+          <table class="table">
+              <thead>
+                  <tr>
+                      <th>Usuario</th>
+                      <th>Creación</th>
+                      <th>Cant.prod/th>
+                      <th>Acciones</th>
+                  </tr>
+              </thead>
+          </table>
+      </div>
+  </div>
+`;
+
+card.innerHTML = contenido;
+
+return card;
